+++ apk/Changelog.md
@@ -7,10 +7,18 @@ The App is divided in two part, **Apk** and **Web** you can check it version in 
 
 The app has a building update check function, so whenever there is a update a red text will show up bellow the main pages clock, also a short warning will pop on the screen let you know if is a **Apk** or **Web** update.
 
-Web Version September 18 2020
+Apk Version 3.0.249 - Web Version September 18 2020
 ==============
 
 * Add extra quality options for the "default player quality" in settings (1080p60, 720p60, etc... etc)
+* Update the player API (ExoPlayer) to latest version, if anyone has a player issues inform, contact information on the about of the app
+* Allow to use the player progressbar on live streams, Twitch keeps only 28 to 32 seconds of it live stream duration on the server, so isn't' possible to go back to the beginning of a streams
+* The above change add a few features requested before see bellow
+* Manually sync multiple streams using the progress bar, on Picture in Picture, 50/50 mode or on Multistream mode, the progress bar will control the main player, on PP mode that is the big player, on 50/50 is the top, on Multistream is the top left or the big one, this way use the progress bar to delay one of the streams until they sink
+* Quick replay using progress bar, use the progress bar to quickly replay a few seconds
+* Lower the latency to the streamer as lower as possible using the progress bar, be aware this will cause re-buffer if the value is too much close to the duration
+* Clips and Live progress bar step is 1 second, allowing more control over the time position
+* General performance improves and bug fixes
 
 Web Version September 17 2020
 ==============